Extract rectangle management setting actions
